Adding SSM Provider
Adding AWS SSM Provider
Adding logic for parsing parameterized FF's
Adding an example filter

+++ ConfigurationProviders/RedisFeatureFlagsProvider.cs
@@ -12,23 +12,22 @@ namespace FeatureManagementSandbox.ConfigurationProviders
     {
         private const string ConnectionString = "localhost";
         private const string SearchPattern = "RedisFeature*";
-        private const ushort ReloadPeriodMs = 10000;
+        private const int ReloadPeriodMs = 10000;
         private static ConnectionMultiplexer _redisClient;
 
         public RedisFeatureFlagsProvider()
         {
             InitializeRedis(ConnectionString);
             ChangeToken.OnChange(() =>
-                {
-                    var cancellationTokenSource = new CancellationTokenSource(ReloadPeriodMs);
-                    var cancellationChangeToken = new CancellationChangeToken(cancellationTokenSource.Token);
-                    return cancellationChangeToken;
-                }, Load);
+            {
+                var cancellationTokenSource = new CancellationTokenSource(ReloadPeriodMs);
+                var cancellationChangeToken = new CancellationChangeToken(cancellationTokenSource.Token);
+                return cancellationChangeToken;
+            }, Load);
         }
         public override void Load()
         {
             var allFeatureFlags = GetAllKeyValuePairs(_redisClient, ConnectionString, SearchPattern);
-
             if (!Data.ContentEquals(allFeatureFlags))
             {
                 Data = allFeatureFlags;
@@ -41,11 +40,11 @@ namespace FeatureManagementSandbox.ConfigurationProviders
             _redisClient = ConnectionMultiplexer.Connect(server);
         }
 
-        private static Dictionary<string,string> GetAllKeyValuePairs(IConnectionMultiplexer host, string hostAddress, string pattern = null)
+        private static Dictionary<string, string> GetAllKeyValuePairs(IConnectionMultiplexer host, string hostAddress, string pattern = null)
         {
             if (string.IsNullOrEmpty(hostAddress)) throw new ArgumentNullException(nameof(hostAddress));
             var db = host.GetDatabase();
-            var keys = host.GetServer(hostAddress, 6379).Keys(pattern:pattern);
+            var keys = host.GetServer(hostAddress, 6379).Keys(pattern: pattern);
             var keysArr = keys.Select(key => (string)key).ToArray();
             var keyValueMap = keysArr.ToDictionary(key => key, key => db.StringGet(key).ToString());
             return keyValueMap;
